Format strings fixes in test suite.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@64711 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ tests/strings/vararg.cpp
@@ -177,9 +177,9 @@ void VarArgTestCase::RepeatedPrintf()
     *p = 'i';
 
     wxString s;
-    s = wxString::Format("buffer %s, len %d", buffer, wxStrlen(buffer));
+    s = wxString::Format("buffer %s, len %d", buffer, (int)wxStrlen(buffer));
     CPPUNIT_ASSERT_EQUAL("buffer hi, len 2", s);
 
-    s = wxString::Format("buffer %s, len %d", buffer, wxStrlen(buffer));
+    s = wxString::Format("buffer %s, len %d", buffer, (int)wxStrlen(buffer));
     CPPUNIT_ASSERT_EQUAL("buffer hi, len 2", s);
 }
